requirements.txt
# Core scientific computing
numpy>=1.21.0,<2.0.0
scipy>=1.7.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration and utilities
pyyaml>=5.4.0
tqdm>=4.62.0

# Statistical analysis
statsmodels>=0.13.0
scikit-learn>=1.0.0

# Optimization (for LP-based core)
pulp>=2.5.0

# Parallel processing
joblib>=1.1.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-xdist>=2.5.0

# Code quality
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Jupyter (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Blockchain integration
web3>=6.0.0
eth-account>=0.8.0
python-dotenv>=0.19.0
click>=8.0.0
rich>=12.0.0

# Solidity compilation (optional for deployment)
py-solc-x>=1.1.0